Created Runelite-specific utils file
Extracted all Runelite-specific properties and functions into a new file.
